requirements.txt
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
idna==3.11
pycparser==3.0
PyGithub==2.8.1
PyJWT==2.11.0
PyNaCl==1.6.2
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.3
